# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse:
    properties:
      code:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode'
      error:
        type: string
    type: object
...
  github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode:
    enum:
    - INVALID_CREDENTIALS
    - UNAUTHORIZED
    - FORBIDDEN
    - EMAIL_NOT_VERIFIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_ALREADY_EXISTS
    - INVALID_REQUEST
    - VALIDATION_FAILED
    - MISSING_REQUIRED_FIELD
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - OPERATION_FAILED
    - EMAIL_IS_ALREADY_IN_USE
    - EMAIL_IS_ALREADY_VERIFIED
    type: string
    x-enum-varnames:
    - ErrCodeInvalidCredentials
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeEmailNotVerified
    - ErrCodeResourceNotFound
    - ErrCodeResourceAlreadyExists
    - ErrCodeInvalidRequest
    - ErrCodeValidationFailed
    - ErrCodeMissingRequiredField
    - ErrCodeInternalServer
    - ErrCodeDatabaseError
    - ErrCodeOperationFailed
    - ErrCodeEmailIsAlreadyInUse
    - ErrCodeEmailIsAlreadyVerified
...
  github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginationResponse'
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.PaginationResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kotolino_lawyer_internal_handlers_responses.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      has_new_appointment:
        type: boolean
      id:
        type: integer
      last_name:
        type: string
      nickname:
        type: string
      profile_image:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.ArticleResponse:
    properties:
      author:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.UserResponse'
      author_id:
        type: integer
      category:
        type: string
      content:
        type: string
      id:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.CreateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    required:
    - category
    - content
    - slug
    - status
    - title
    type: object
...
  internal_handlers.UpdateArticleRequest:
    properties:
      category:
        type: string
      content:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /articles:
    get:
      description: Retrieves a paginated list of articles with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: params[category]
        type: string
      - description: Search in title
        in: query
        name: params[query]
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handlers.ArticleResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get all articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      description: Creates a new article
      parameters:
      - description: Article data
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.ArticleResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Create new article
      tags:
      - articles
  /articles/{id}:
    delete:
      description: Deletes article by ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Article deleted
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid article ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Delete article
      tags:
      - articles
    get:
      description: Returns the article details by ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ArticleResponse'
        "400":
          description: Invalid article ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get article by ID
      tags:
      - articles
    put:
      consumes:
      - application/json
      description: Updates an existing article by ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated article data
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ArticleResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Update article
      tags:
      - articles
  /articles/slug/{slug}:
    get:
      description: Returns the article details by slug
      parameters:
      - description: Article Slug
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ArticleResponse'
        "404":
          description: Article not found
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      summary: Get article by slug
      tags:
      - articles
...
swagger: "2.0"